requirement.txt
beautifulsoup4
requests
Pillow
tqdm
